Changement structure bot
Changements handlers, modification structure

+++ commands/nick.js
@@ -1,11 +1,20 @@
 const Discord = require('discord.js');
 const fs = require('fs');
-
-module.exports = function (message, client) {
+const Command = require("../utils/commandHandler.js");
+module.exports = class Botinfo extends Command {
+	constructor() {
+		super({
+			name: "nick",
+			category: "admin",
+			aliases: [],
+			description: "Commande permettant d'enlever les carachtères spéciaux d'un pseudo",
+			usage: "{{prefix}}nick {Membre"
+		});
+	}
+run(message, client) {
         const membre = message.mentions.members.first()
         const role = message.guild.roles.cache.find(role => role.id === '791681762124103721')
-        const role2 = message.guild.roles.cache.find(role => role.id === '792889476091215892')  
-        //if (!(message.member.roles.cache.has(role) || message.author.id === '398126558432329728' || message.member.roles.cache.has(role2))) return message.channel.send("Vous n'avez pas la permission");
+        if (!(message.member.roles.cache.has(role) || message.author.id === '398126558432329728')) return message.channel.send("Vous n'avez pas la permission");
         if(!membre) {
             return message.channel.send(`Veuillez mentionner la personne à qui enlevé les charactères spéciaux.`).then(msg => msg.delete({timeout: 4000})).then(message.delete());
         }
@@ -25,3 +34,4 @@ module.exports = function (message, client) {
         membre.setNickname(m).catch(console.error)
         message.channel.send(`Le pseudo de ${membre} a bien été enlevé des charactères spéciaux`)
 }
+};
